comarch_klask_league/urls.py
"""comarch_klask_league URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/1.10/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.conf.urls import url, include
    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
"""
from django.conf import settings
from django.conf.urls import include, url, static
from django.contrib import admin, auth
from django.contrib.auth import views
from django.urls import path

from . import settings, views

urlpatterns = [
    url(r'^$', views.index, name='home'),
    url(r'^new_player$', views.new_player, name='new_player'),
    url(r'^new_match$', views.new_match, name='new_match'),
    url(r'^player_history$', views.player_history, name='player_history'),
    path('admin/', admin.site.urls),
]

if settings.DEBUG:
    import debug_toolbar
    urlpatterns = [
        path('__debug__/', include(debug_toolbar.urls)),

    ] + urlpatterns